requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.5
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
pydantic==2.10.2
pydantic_core==2.27.1
pydub==0.25.1
python-multipart==0.0.17
regex==2024.11.6
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.11.0
starlette==0.41.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
